Supix WG2: Simple PianoKeyboard demo now works + latest WG3.

+++ framework/gfxdevices/stream/wg3_gfxstreamplayer.cpp
@@ -379,7 +379,7 @@ namespace wg
 		{
 			CoordI		begin;
 			Direction	dir;
-			uint16_t	length;
+			int			length;
 			HiColor		color;
 			float		thickness;
